Fix, refator and cleanup tests
